B4A 12.8, ExoPlayer 3.0

+++ Libs_XUI/src/anywheresoftware/b4a/objects/B4XCanvas.java
@@ -156,7 +156,7 @@ public class B4XCanvas {
 		public float getWidth() {
 			return rf.right - rf.left;
 		}
-		public void setWidth(int w) {
+		public void setWidth(float w) {
 			rf.right = rf.left + w;
 		}
 		/**
